Moved CGFloat & UIColor extensions
